# swagger.yaml
openapi: 3.0.3
info:
  title: Eco Tips - OpenAPI 3.0
  description: |-
    Eco Tips is a web application that allows its users, in a fun way, to reduce their ecological impact and save money.

    In the current context, more and more citizens are aware of the climate issues and need to take action in their daily lives to reduce their carbon footprint. The country's economic context also leads citizens to reduce their daily expenses. To meet this need, Eco Tips offers its users eco-gestures based in part on a [study by CLER - Network for Energy Transition](https://cler.org/wp-content/uploads/2021/01/Guide-100-e%CC%81cogestes-De%CC%81clics_planches.pdf), helping them to save money and live more sustainably.

    Therefore, Eco Tips provides a fun solution that encourages individuals to awaken their environmental awareness, to sustainably install eco-gestures in their daily lives in order to reduce the users' carbon footprint. The application informs users about new uses while creating a community of committed citizens.
    
    Some useful links:
    - [Eco Tips back Github repository](https://github.com/O-clock-Ohm/projet-03-eco-tips-back)
    - [Eco Tips front Github repository](https://github.com/O-clock-Ohm/projet-03-eco-tips-front)
servers:
  - url : http://tonympt-server.eddi.cloud:8080/
...
paths:
...
  /me/user:
    get:
      tags:
        - user
      summary: Returns the user's data
      description: Returns the user's data hiding their password
      operationId: getProfile
      responses:
        '200':
          description: Successful operation
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Veuillez vous authentifier pour accéder à cette page/Vous n'êtes pas autorisé à accéder à cette page.
      security:
        - Jwt: []
    patch:
      tags:
        - user
      summary: Updates the user's data (password not included)
      description: Updates one to many user's data
      operationId: updateProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                firstname:
                  type: string
                  description: user's firstname
                  example: John
                lastname:
                  type: string
                  description: user's lastname
                  example: Doe
                email:
                  type: string
                  description: user's email
                  pattern: '^[a-zA-Z0-9.!#$%&’*+/=?^_`{|}~-]+@[a-zA-Z0-9-]+(?:\.[a-zA-Z0-9-]+)*$'
                  example: john.doe@example.com
                birthdate:
                  type: string
                  description: user's date of birth
                  example: 1970-01-01
      responses:
        '204':
          description: Successful operation
        '401':
          description: Invalid identifiers/Inputs must not be empty
      security:
        - Jwt: []
    delete:
      tags:
        - user
      summary: Deletes the user profile
      description: Deletes user from the app
      operationId: deleteProfile
      responses:
        '204':
          description: Successful operation
      security:
        - Jwt: []
  /me/user/password:
    patch:
      tags:
        - user
      summary: Updates user's password
      description: Updates user's password
      operationId: updatePassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                password:
                  type: string
                  description: user's new password
                  pattern: /^(?=.*?[A-Z])(?=.*?[a-z])(?=.*?[0-9])(?=.*?[#?!@$%^&*-]).{8,}$/
                  example: Azertyuiop*4
                confirmpassword:
                  type: string
                  description: new password confirmation
                  example: Azertyuiop*4
      responses:
        '204':
          description: Successful operation
        '401':
          description: Invalid identifiers/Inputs must not be empty
          content:
            application/json:
              schema:
                type: object
                properties:
                    message:
                      type: string
                      example: 
                        - "Merci de renseigner un mot de passe différent du précédent."
                        - "Couple login/mot de passe est incorrect."
                        - "Format du mot de passe incorrect : 8 caractères minimum comprenant au minimum 1 chiffre, 1 lettre miniscule, 1 lettre majuscule et 1 caractère spécial."
                        - "Le champ 'nom du champ' ne peut pas être vide."
  /user:
    get:
      tags:
        - user
      summary: Returns all users' data (admin only)
      description: Returns all users' data
      operationId: getAllUsers
      responses:
        '200':
          description: Successful operation
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Veuillez vous authentifier pour accéder à cette page/Vous n'êtes pas autorisé à accéder à cette page.
      security:
        - Jwt: []
  /user/{id}:
    patch:
      tags:
        - user
      summary: Sets user's role to admin (admin only)
      description: Sets user's role to admin
      operationId: updateUserAsAdmin
      responses:
        '204':
          description: Successful operation
        '401':
          description: Veuillez vous authentifier pour accéder à cette page/Vous n'êtes pas autorisé à accéder à cette page.
      security:
        - Jwt: []
...
  /ranking/score:
    get:
      tags:
        - ranking
      summary: Returns the 5 best users ordered by score
      description: Returns the 5 best users ordered by score
      operationId: getRankingScore
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    user:
                      type: string
                      description: user's firstname and last name
                      example: John Doe
                    score:
                      type: integer
                      example: 60
                minItems: 5
                maxItems: 5
        '401':
          description: Veuillez vous authentifier pour accéder à cette page/Vous n'êtes pas autorisé à accéder à cette page.
      security:
        - Jwt: []
  /ranking/creation:
    get:
      tags:
        - ranking
      summary: Returns the 5 best users ordered by their card creation amount
      description: Returns the 5 best users ordered by their card creation amount
      operationId: getRankingCreation
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    user:
                      type: string
                      description: user's firstname and last name
                      example: John Doe
                    cards_created:
                      type: integer
                      example: 3
                minItems: 5
                maxItems: 5
        '401':
          description: Veuillez vous authentifier pour accéder à cette page/Vous n'êtes pas autorisé à accéder à cette page.
      security:
        - Jwt: []
...
components:
  schemas:
...
    User:
      type: object
      properties:
        firstname:
          type: string
          example: Laura
        lastname:
          type: string
          example: Teur
        email:
          type: string
          format: email
          example: user@example.com
        password: 
          type: string
          description: >
            Password returned will be replaced with '*'.
          example: '***********'
        birthdate: 
          type: string
          format: date
        ecocoins: 
          type: int
          description: The user's number of ecocoins which allow them to get cards
        score: 
          type: int
          format: The user's score obtained by validating cards
      xml:
        name: user
...